Fix/ignore clippy warning
Allow the following warnings:

- clippy::needless_range_loop
- clippy::single_match
- clippy::uninlined_format_args

Fix the following warnings:

- clippy::borrow_deref_ref
- clippy::should_implement_trait
- clippy::unnecessary_cast

+++ src/app/mapper.rs
@@ -8,7 +8,7 @@ impl Mapper for UefiMapper {
         address: PhysicalAddress,
         _size: usize,
     ) -> Result<VirtualAddress, &'static str> {
-        Ok(VirtualAddress(address.0 as usize))
+        Ok(VirtualAddress(address.0))
     }
 
     unsafe fn unmap_aligned(
